fix: Select width 0px when searching
